Updated the sound playback.
There seems to be a fundemental problem in almost all Java libraries
which stops multi-channel wave files playing. This is also an issue in
JavaFX MediaPlayer etc. Have created a new function using the Java Sound
API which only plays the first channel bytes.

Needs to be threaded so non functional just now.